+++ packages/danmaku-anywhere/src/common/components/image/useImage.tsx
@@ -1,6 +1,6 @@
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
+import { fetchImageBase64, getOrFetchCachedImage } from '@/common/image/cache'
 import { imageQueryKeys } from '@/common/queries/queryKeys'
-import { getOrFetchCachedImage, fetchImageBase64 } from '@/common/image/cache'
 
 /**
  * Fetching images from the content script has issues with the same-origin policy and images in the extension assets.
